pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.0",
    "httpx-sse>=0.4.0",
    "streamlit>=1.45.0",
    "google-genai>=1.68.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.1",
    "requests>=2.33.0",
    "sqlalchemy>=2.0.48",
    "sse-starlette>=2.1.3",
    "uvicorn>=0.42.0",
    "voyageai>=0.3.7",
]

[tool.uv]
dev-dependencies = ["pytest>=8.0.0"]
